pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = "oAuth test"
authors = ["plat01 <fplatoff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.32.0"
motor = "^3.6.0"
fastapi = "^0.115.4"
pydantic-settings = "^2.6.1"
beanie = "^1.27.0"
email-validator = "^2.2.0"
google-auth-oauthlib = "^1.2.1"
google-api-python-client = "^2.154.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
